Fix some merge conflicts

+++ src/main/java/mekanism/common/item/ItemNetworkReader.java
@@ -79,7 +79,7 @@ public class ItemNetworkReader extends ItemEnergized {
                     }
                     CapabilityUtils.getCapability(tileEntity, Capabilities.HEAT_HANDLER_CAPABILITY, opposite).ifPresent(heatHandler ->
                           player.sendMessage(MekanismLang.NETWORK_READER_ABOVE_AMBIENT.translateColored(EnumColor.GRAY, EnumColor.DARK_GRAY,
-                                MekanismUtils.getTemperatureDisplay(heatHandler.getTotalTemperature().doubleValue(), TemperatureUnit.KELVIN))));
+                                MekanismUtils.getTemperatureDisplay(heatHandler.getTotalTemperature(), TemperatureUnit.KELVIN))));
                     player.sendMessage(MekanismLang.NETWORK_READER_BORDER.translateColored(EnumColor.GRAY, "-------------", EnumColor.DARK_BLUE, "[=======]"));
                 } else {
                     Optional<IHeatHandler> heatHandler = MekanismUtils.toOptional(CapabilityUtils.getCapability(tileEntity, Capabilities.HEAT_HANDLER_CAPABILITY, opposite));
@@ -87,7 +87,7 @@ public class ItemNetworkReader extends ItemEnergized {
                         IHeatHandler transfer = heatHandler.get();
                         MekanismLang.NETWORK_READER_BORDER.translateColored(EnumColor.GRAY, "-------------",
                               MekanismLang.GENERIC_SQUARE_BRACKET.translateColored(EnumColor.DARK_BLUE, MekanismLang.MEKANISM));
-                        ITextComponent temp = MekanismUtils.getTemperatureDisplay(transfer.getTotalTemperature().doubleValue(), TemperatureUnit.KELVIN);
+                        ITextComponent temp = MekanismUtils.getTemperatureDisplay(transfer.getTotalTemperature(), TemperatureUnit.KELVIN);
                         player.sendMessage(MekanismLang.NETWORK_READER_ABOVE_AMBIENT.translateColored(EnumColor.GRAY, EnumColor.DARK_GRAY, temp));
                         player.sendMessage(MekanismLang.NETWORK_READER_BORDER.translateColored(EnumColor.GRAY, "-------------", EnumColor.DARK_BLUE, "[=======]"));
                     } else {
